refactor: remove mapping from all project (except handlers))
