+++ Source/Application/Kysect.Shreks.Application.Commands/Commands/CreateSubmissionCommand.cs
@@ -1,6 +1,6 @@
 ﻿using CommandLine;
 
-namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+namespace Kysect.Shreks.Application.Commands.Commands;
 
 [Verb("/create-submission")]
 public class CreateSubmissionCommand : IShreksCommand
