+++ Source/Application/Kysect.Shreks.Application.Commands/Commands/IShreksCommand.cs
@@ -1,0 +1,5 @@
+namespace Kysect.Shreks.Application.Commands.Commands;
+
+public interface IShreksCommand
+{
+}
