+++ Source/Application/Kysect.Shreks.Application.Commands/Commands/RateCommand.cs
@@ -1,6 +1,6 @@
 using CommandLine;
 
-namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
+namespace Kysect.Shreks.Application.Commands.Commands;
 
 [Verb("/rate", aliases: new []{"/assess"})]
 public class RateCommand : IShreksCommand
