+++ Source/Application/Kysect.Shreks.Application.Commands/Contexts/SubmissionContext.cs
@@ -1,13 +1,14 @@
 using Kysect.Shreks.Application.Dto.Study;
+using Kysect.Shreks.Core.Submissions;
 using Microsoft.Extensions.Logging;
 
 namespace Kysect.Shreks.Application.Commands.Contexts;
 
 public class SubmissionContext : BaseContext
 {
-    public SubmissionDto Submission { get; }
+    public Submission Submission { get; }
 
-    public SubmissionContext(ILogger log, Guid issuerId, SubmissionDto submission) 
+    public SubmissionContext(ILogger log, Guid issuerId, Submission submission) 
         : base(log, issuerId)
     {
         Submission = submission;
