--- Source/Application/Kysect.Shreks.Application.UserCommands.Abstractions/Commands/IShreksCommand.cs
@@ -1,5 +1,0 @@
-namespace Kysect.Shreks.Application.UserCommands.Abstractions.Commands;
-
-public interface IShreksCommand
-{
-}
