+++ Source/Kysect.Shreks.sln
@@ -63,8 +63,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Application.G
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Kysect.Shreks.Application.GithubWorkflow.Abstractions", "Application\Kysect.Shreks.Application.GithubWorkflow.Abstractions\Kysect.Shreks.Application.GithubWorkflow.Abstractions.csproj", "{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Kysect.Shreks.Application.UserCommands.Abstractions", "Application\Kysect.Shreks.Application.UserCommands.Abstractions\Kysect.Shreks.Application.UserCommands.Abstractions.csproj", "{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -163,10 +161,6 @@ Global
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68}.Release|Any CPU.Build.0 = Release|Any CPU
-		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -196,7 +190,6 @@ Global
 		{601BCCAB-8895-43CB-BC46-EE2C20F4031F} = {DC866D57-D41E-455D-BD7D-85F1DD575823}
 		{DF0C2310-8F92-404C-885E-D495B3F334EB} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 		{3B0EB40E-6E3C-4F8D-BBD7-9B7FC5FB7D68} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
-		{3B01FC98-191F-4CF0-9709-D7EFCBC889D0} = {AE556858-2A9D-481F-9235-3C74F0CDC5A5}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {CD5C67DB-70F8-4C47-A3F4-7C94BECE1657}
